+++ src/tweens/builders/GetEaseFunction.js
@@ -5,17 +5,20 @@
  */
 
 var EaseMap = require('../../math/easing/EaseMap');
+var UppercaseFirst = require('../../utils/string/UppercaseFirst');
 
 /**
- * [description]
+ * This internal function is used to return the correct ease function for a Tween.
+ * 
+ * It can take a variety of input, including an EaseMap based string, or a custom function.
  *
  * @function Phaser.Tweens.Builders.GetEaseFunction
  * @since 3.0.0
  *
- * @param {(string|function)} ease - [description]
- * @param {array} easeParams - [description]
+ * @param {(string|function)} ease - The ease to find. This can be either a string from the EaseMap, or a custom function.
+ * @param {number[]} [easeParams] - An optional array of ease parameters to go with the ease.
  *
- * @return {function} [description]
+ * @return {function} The ease function.
  */
 var GetEaseFunction = function (ease, easeParams)
 {
@@ -23,10 +26,49 @@ var GetEaseFunction = function (ease, easeParams)
     var easeFunction = EaseMap.Power0;
 
     //  Prepare ease function
-    if (typeof ease === 'string' && EaseMap.hasOwnProperty(ease))
+    if (typeof ease === 'string')
     {
         //  String based look-up
-        easeFunction = EaseMap[ease];
+
+        //  1) They specified it correctly
+        if (EaseMap.hasOwnProperty(ease))
+        {
+            easeFunction = EaseMap[ease];
+        }
+        else
+        {
+            //  Do some string manipulation to try and find it
+            var direction = '';
+
+            if (ease.indexOf('.'))
+            {
+                //  quad.in = Quad.easeIn
+                //  quad.out = Quad.easeOut
+                //  quad.inout =Quad.easeInOut
+
+                direction = ease.substr(ease.indexOf('.') + 1);
+
+                if (direction.toLowerCase() === 'in')
+                {
+                    direction = 'easeIn';
+                }
+                else if (direction.toLowerCase() === 'out')
+                {
+                    direction = 'easeOut';
+                }
+                else if (direction.toLowerCase() === 'inout')
+                {
+                    direction = 'easeInOut';
+                }
+            }
+
+            ease = UppercaseFirst(ease.substr(0, ease.indexOf('.') + 1) + direction);
+
+            if (EaseMap.hasOwnProperty(ease))
+            {
+                easeFunction = EaseMap[ease];
+            }
+        }
     }
     else if (typeof ease === 'function')
     {