+++ src/tweens/tween/Tween.js
@@ -161,9 +161,22 @@ var Tween = new Class({
          *
          * @name Phaser.Tweens.Tween#hasStarted
          * @type {boolean}
+         * @readonly
          * @since 3.19.0
          */
         this.hasStarted = false;
+
+        /**
+         * Is this Tween currently seeking?
+         * This boolean is toggled in the `Tween.seek` method.
+         * When a tween is seeking it will not dispatch any events or callbacks.
+         *
+         * @name Phaser.Tweens.Tween#isSeeking
+         * @type {boolean}
+         * @readonly
+         * @since 3.19.0
+         */
+        this.isSeeking = false;
 
         /**
          * Time in ms/frames before the 'onComplete' event fires. This never fires if loop = -1 (as it never completes)
@@ -295,11 +308,20 @@ var Tween = new Class({
         this.totalProgress = 0;
 
         /**
-         * An object containing the various Tween callback references.
+         * An object containing the different Tween callback functions.
+         * 
+         * You can either set these in the Tween config, or by calling the `Tween.setCallback` method.
+         * 
+         * `onActive` When the Tween is moved from the pending to the active list in the Tween Manager, even if playback paused.
+         * `onStart` When the Tween starts playing after a delayed state. Will happen at the same time as `onActive` if it has no delay.
+         * `onYoyo` When a TweenData starts a yoyo. This happens _after_ the `hold` delay expires, if set.
+         * `onRepeat` When a TweenData repeats playback. This happens _after_ the `repeatDelay` expires, if set.
+         * `onComplete` When the Tween finishes playback fully or `Tween.stop` is called. Never invoked if tween is set to repeat infinitely.
+         * `onUpdate` When a TweenData updates a property on a source target during playback.
+         * `onLoop` When a Tween loops. This happens _after_ the `loopDelay` expires, if set.
          *
          * @name Phaser.Tweens.Tween#callbacks
          * @type {object}
-         * @default 0
          * @since 3.0.0
          */
         this.callbacks = {
@@ -317,7 +339,6 @@ var Tween = new Class({
          *
          * @name Phaser.Tweens.Tween#callbackScope
          * @type {any}
-         * @default 0
          * @since 3.0.0
          */
         this.callbackScope;
@@ -584,12 +605,17 @@ var Tween = new Class({
             var tweenData = data[i];
             var target = tweenData.target;
             var gen = tweenData.gen;
+            var key = tweenData.key;
+            var targetIndex = tweenData.index;
 
-            tweenData.delay = gen.delay(i, totalTargets, target);
-            tweenData.duration = Math.max(gen.duration(i, totalTargets, target), 0.001);
-            tweenData.hold = gen.hold(i, totalTargets, target);
-            tweenData.repeat = gen.repeat(i, totalTargets, target);
-            tweenData.repeatDelay = gen.repeatDelay(i, totalTargets, target);
+            //  Old function signature: i, totalTargets, target
+            //  New function signature: target, key, value, index, total, tween
+
+            tweenData.delay = gen.delay(target, key, 0, targetIndex, totalTargets, this);
+            tweenData.duration = Math.max(gen.duration(target, key, 0, targetIndex, totalTargets, this), 0.001);
+            tweenData.hold = gen.hold(target, key, 0, targetIndex, totalTargets, this);
+            tweenData.repeat = gen.repeat(target, key, 0, targetIndex, totalTargets, this);
+            tweenData.repeatDelay = gen.repeatDelay(target, key, 0, targetIndex, totalTargets, this);
         }
 
         this.calcDuration();
@@ -616,25 +642,15 @@ var Tween = new Class({
     {
         this.parent.makeActive(this);
 
-        //  When the Tween is moved from the pending to the active list in the manager, even if playback delayed
-        this.emit(Events.TWEEN_ACTIVE, this);
-
-        var onActive = this.callbacks.onActive;
-
-        if (onActive)
-        {
-            onActive.params[1] = this.targets;
-
-            onActive.func.apply(onActive.scope, onActive.params);
-        }
+        this.dispatchTweenEvent(Events.TWEEN_ACTIVE, this.callbacks.onActive);
     },
 
     /**
      * Internal method that advances to the next state of the Tween during playback.
      *
      * @method Phaser.Tweens.Tween#nextState
-     * @fires Phaser.Tweens.Events#TWEEN_LOOP
      * @fires Phaser.Tweens.Events#TWEEN_COMPLETE
+     * @fires Phaser.Tweens.Events#TWEEN_LOOP
      * @since 3.0.0
      */
     nextState: function ()
@@ -654,39 +670,22 @@ var Tween = new Class({
             }
             else
             {
-                this.emit(Events.TWEEN_LOOP, this, this.targets);
-
-                var onLoop = this.callbacks.onLoop;
-    
-                if (onLoop)
-                {
-                    onLoop.params[1] = this.targets;
-    
-                    onLoop.func.apply(onLoop.scope, onLoop.params);
-                }
-
                 this.state = TWEEN_CONST.ACTIVE;
+
+                this.dispatchTweenEvent(Events.TWEEN_LOOP, this.callbacks.onLoop);
             }
         }
         else if (this.completeDelay > 0)
         {
-            this.countdown = this.completeDelay;
             this.state = TWEEN_CONST.COMPLETE_DELAY;
+
+            this.countdown = this.completeDelay;
         }
         else
         {
-            this.emit(Events.TWEEN_COMPLETE, this, this.targets);
-
-            var onComplete = this.callbacks.onComplete;
-
-            if (onComplete)
-            {
-                onComplete.params[1] = this.targets;
-
-                onComplete.func.apply(onComplete.scope, onComplete.params);
-            }
-
             this.state = TWEEN_CONST.PENDING_REMOVE;
+
+            this.dispatchTweenEvent(Events.TWEEN_COMPLETE, this.callbacks.onComplete);
         }
     },
 
@@ -802,10 +801,16 @@ var Tween = new Class({
     resetTweenData: function (resetFromLoop)
     {
         var data = this.data;
+        var total = this.totalData;
+        var totalTargets = this.totalTargets;
 
-        for (var i = 0; i < this.totalData; i++)
+        for (var i = 0; i < total; i++)
         {
             var tweenData = data[i];
+
+            var target = tweenData.target;
+            var key = tweenData.key;
+            var targetIndex = tweenData.index;
 
             tweenData.progress = 0;
             tweenData.elapsed = 0;
@@ -814,9 +819,9 @@ var Tween = new Class({
 
             if (resetFromLoop)
             {
-                tweenData.start = tweenData.getStartValue(tweenData.target, tweenData.key, tweenData.start);
+                tweenData.start = tweenData.getStartValue(target, key, tweenData.start, targetIndex, totalTargets, this);
 
-                tweenData.end = tweenData.getEndValue(tweenData.target, tweenData.key, tweenData.end);
+                tweenData.end = tweenData.getEndValue(target, key, tweenData.end, targetIndex, totalTargets, this);
 
                 tweenData.current = tweenData.start;
 
@@ -830,7 +835,13 @@ var Tween = new Class({
             if (tweenData.delay > 0)
             {
                 tweenData.elapsed = tweenData.delay;
+
                 tweenData.state = TWEEN_CONST.DELAY;
+            }
+
+            if (tweenData.getActiveValue)
+            {
+                target[key] = tweenData.getActiveValue(tweenData.target, tweenData.key, tweenData.start);
             }
         }
     },
@@ -888,7 +899,7 @@ var Tween = new Class({
 
         if (this.totalDuration >= 3600000)
         {
-            console.warn('Cannot Tween.seek - duration too long or infinite repeat');
+            console.warn('Tween.seek duration too long');
 
             return this;
         }
@@ -911,22 +922,27 @@ var Tween = new Class({
             var tweenData = data[i];
             var target = tweenData.target;
             var gen = tweenData.gen;
+            var key = tweenData.key;
+            var targetIndex = tweenData.index;
 
             tweenData.progress = 0;
             tweenData.elapsed = 0;
 
             tweenData.repeatCounter = (tweenData.repeat === -1) ? 999999999999 : tweenData.repeat;
 
-            tweenData.delay = gen.delay(i, totalTargets, target);
-            tweenData.duration = Math.max(gen.duration(i, totalTargets, target), 0.001);
-            tweenData.hold = gen.hold(i, totalTargets, target);
-            tweenData.repeat = gen.repeat(i, totalTargets, target);
-            tweenData.repeatDelay = gen.repeatDelay(i, totalTargets, target);
+            //  Old function signature: i, totalTargets, target
+            //  New function signature: target, key, value, index, total, tween
+
+            tweenData.delay = gen.delay(target, key, 0, targetIndex, totalTargets, this);
+            tweenData.duration = Math.max(gen.duration(target, key, 0, targetIndex, totalTargets, this), 0.001);
+            tweenData.hold = gen.hold(target, key, 0, targetIndex, totalTargets, this);
+            tweenData.repeat = gen.repeat(target, key, 0, targetIndex, totalTargets, this);
+            tweenData.repeatDelay = gen.repeatDelay(target, key, 0, targetIndex, totalTargets, this);
 
             tweenData.current = tweenData.start;
             tweenData.state = TWEEN_CONST.PLAYING_FORWARD;
 
-            this.updateTweenData(this, tweenData, 0);
+            this.updateTweenData(this, tweenData, 0, targetIndex, totalTargets);
 
             if (tweenData.delay > 0)
             {
@@ -946,11 +962,15 @@ var Tween = new Class({
             this.state = TWEEN_CONST.ACTIVE;
         }
 
+        this.isSeeking = true;
+
         do
         {
             this.update(0, delta);
 
         } while (this.totalProgress < toPosition);
+
+        this.isSeeking = false;
 
         if (wasPaused)
         {
@@ -963,17 +983,23 @@ var Tween = new Class({
     /**
      * Sets an event based callback to be invoked during playback.
      * 
-     * The `type` can be: 'onActive', 'onComplete', 'onLoop', 'onRepeat', 'onStart',
-     * 'onUpdate' or 'onYoyo'.
+     * Calling this method will replace a previously set callback for the given type, if any exists.
      * 
-     * Calling this method will replace a previously set callback for the
-     * given type, if any exists.
+     * The types available are:
+     * 
+     * `onActive` When the Tween is moved from the pending to the active list in the Tween Manager, even if playback paused.
+     * `onStart` When the Tween starts playing after a delayed state. Will happen at the same time as `onActive` if it has no delay.
+     * `onYoyo` When a TweenData starts a yoyo. This happens _after_ the `hold` delay expires, if set.
+     * `onRepeat` When a TweenData repeats playback. This happens _after_ the `repeatDelay` expires, if set.
+     * `onComplete` When the Tween finishes playback fully or `Tween.stop` is called. Never invoked if tween is set to repeat infinitely.
+     * `onUpdate` When a TweenData updates a property on a source target during playback.
+     * `onLoop` When a Tween loops. This happens _after_ the `loopDelay` expires, if set.
      *
      * @method Phaser.Tweens.Tween#setCallback
      * @since 3.0.0
      *
-     * @param {string} type - Type of the callback.
-     * @param {function} callback - Callback function.
+     * @param {string} type - Type of the callback to set.
+     * @param {function} callback - The function to invoke when this callback happens.
      * @param {array} [params] - An array of parameters for specified callbacks types.
      * @param {any} [scope] - The context the callback will be invoked in.
      *
@@ -1008,23 +1034,15 @@ var Tween = new Class({
 
         if (delay)
         {
-            this.countdown = delay;
             this.state = TWEEN_CONST.COMPLETE_DELAY;
+
+            this.countdown = delay;
         }
         else
         {
-            this.emit(Events.TWEEN_COMPLETE, this, this.targets);
-
-            var onComplete = this.callbacks.onComplete;
-
-            if (onComplete)
-            {
-                onComplete.params[1] = this.targets;
-
-                onComplete.func.apply(onComplete.scope, onComplete.params);
-            }
-
             this.state = TWEEN_CONST.PENDING_REMOVE;
+
+            this.dispatchTweenEvent(Events.TWEEN_COMPLETE, this.callbacks.onComplete);
         }
 
         return this;
@@ -1098,8 +1116,8 @@ var Tween = new Class({
      *
      * @method Phaser.Tweens.Tween#update
      * @fires Phaser.Tweens.Events#TWEEN_COMPLETE
-     * @fires Phaser.Tweens.Events#TWEEN_START
      * @fires Phaser.Tweens.Events#TWEEN_LOOP
+     * @fires Phaser.Tweens.Events#TWEEN_START
      * @since 3.0.0
      *
      * @param {number} timestamp - The current time. Either a High Resolution Timer value if it comes from Request Animation Frame, or Date.now if using SetTimeout.
@@ -1131,7 +1149,7 @@ var Tween = new Class({
         {
             case TWEEN_CONST.ACTIVE:
 
-                if (!this.hasStarted)
+                if (!this.hasStarted && !this.isSeeking)
                 {
                     this.startDelay -= delta;
         
@@ -1139,16 +1157,7 @@ var Tween = new Class({
                     {
                         this.hasStarted = true;
 
-                        this.emit(Events.TWEEN_START, this, this.targets);
-
-                        var onStart = this.callbacks.onStart;
-
-                        if (onStart)
-                        {
-                            onStart.params[1] = this.targets;
-
-                            onStart.func.apply(onStart.scope, onStart.params);
-                        }
+                        this.dispatchTweenEvent(Events.TWEEN_START, this.callbacks.onStart);
                     }
                 }
 
@@ -1178,18 +1187,9 @@ var Tween = new Class({
 
                 if (this.countdown <= 0)
                 {
-                    this.emit(Events.TWEEN_LOOP, this, this.targets);
-
-                    var onLoop = this.callbacks.onLoop;
-        
-                    if (onLoop)
-                    {
-                        onLoop.params[1] = this.targets;
-        
-                        onLoop.func.apply(onLoop.scope, onLoop.params);
-                    }
-
                     this.state = TWEEN_CONST.ACTIVE;
+
+                    this.dispatchTweenEvent(Events.TWEEN_LOOP, this.callbacks.onLoop);
                 }
 
                 break;
@@ -1211,22 +1211,64 @@ var Tween = new Class({
 
                 if (this.countdown <= 0)
                 {
-                    this.emit(Events.TWEEN_COMPLETE, this, this.targets);
-
-                    var onComplete = this.callbacks.onComplete;
-
-                    if (onComplete)
-                    {
-                        onComplete.func.apply(onComplete.scope, onComplete.params);
-                    }
-
                     this.state = TWEEN_CONST.PENDING_REMOVE;
+
+                    this.dispatchTweenEvent(Events.TWEEN_COMPLETE, this.callbacks.onComplete);
                 }
 
                 break;
         }
 
         return (this.state === TWEEN_CONST.PENDING_REMOVE);
+    },
+
+    /**
+     * Internal method that will emit a TweenData based Event and invoke the given callback.
+     *
+     * @method Phaser.Tweens.Tween#dispatchTweenDataEvent
+     * @since 3.19.0
+     *
+     * @param {Phaser.Tweens.Events} event - The Event to be dispatched.
+     * @param {function} callback - The callback to be invoked. Can be `null` or `undefined` to skip invocation.
+     * @param {Phaser.Types.Tweens.TweenDataConfig} tweenData - The TweenData object that caused this event.
+     */
+    dispatchTweenDataEvent: function (event, callback, tweenData)
+    {
+        if (!this.isSeeking)
+        {
+            this.emit(event, this, tweenData.key, tweenData.target);
+
+            if (callback)
+            {
+                callback.params[1] = tweenData.target;
+
+                callback.func.apply(callback.scope, callback.params);
+            }
+        }
+    },
+
+    /**
+     * Internal method that will emit a Tween based Event and invoke the given callback.
+     *
+     * @method Phaser.Tweens.Tween#dispatchTweenEvent
+     * @since 3.19.0
+     *
+     * @param {Phaser.Tweens.Events} event - The Event to be dispatched.
+     * @param {function} callback - The callback to be invoked. Can be `null` or `undefined` to skip invocation.
+     */
+    dispatchTweenEvent: function (event, callback)
+    {
+        if (!this.isSeeking)
+        {
+            this.emit(event, this, this.targets);
+
+            if (callback)
+            {
+                callback.params[1] = this.targets;
+    
+                callback.func.apply(callback.scope, callback.params);
+            }
+        }
     },
 
     /**
@@ -1263,19 +1305,9 @@ var Tween = new Class({
                 tweenData.target.toggleFlipY();
             }
 
-            this.emit(Events.TWEEN_YOYO, this, tweenData.key, tweenData.target);
+            this.dispatchTweenDataEvent(Events.TWEEN_YOYO, tween.callbacks.onYoyo, tweenData);
 
-            var onYoyo = tween.callbacks.onYoyo;
-
-            if (onYoyo)
-            {
-                //  Element 1 is reserved for the target of the yoyo (and needs setting here)
-                onYoyo.params[1] = tweenData.target;
-
-                onYoyo.func.apply(onYoyo.scope, onYoyo.params);
-            }
-
-            tweenData.start = tweenData.getStartValue(tweenData.target, tweenData.key, tweenData.start);
+            tweenData.start = tweenData.getStartValue(tweenData.target, tweenData.key, tweenData.start, tweenData.index, tween.totalTargets, tween);
 
             return TWEEN_CONST.PLAYING_BACKWARD;
         }
@@ -1300,9 +1332,9 @@ var Tween = new Class({
                 tweenData.target.toggleFlipY();
             }
 
-            tweenData.start = tweenData.getStartValue(tweenData.target, tweenData.key, tweenData.start);
+            tweenData.start = tweenData.getStartValue(tweenData.target, tweenData.key, tweenData.start, tweenData.index, tween.totalTargets, tween);
 
-            tweenData.end = tweenData.getEndValue(tweenData.target, tweenData.key, tweenData.start);
+            tweenData.end = tweenData.getEndValue(tweenData.target, tweenData.key, tweenData.start, tweenData.index, tween.totalTargets, tween);
 
             //  Delay?
             if (tweenData.repeatDelay > 0)
@@ -1317,17 +1349,7 @@ var Tween = new Class({
             }
             else
             {
-                this.emit(Events.TWEEN_REPEAT, this, tweenData.key, tweenData.target);
-
-                var onRepeat = tween.callbacks.onRepeat;
-    
-                if (onRepeat)
-                {
-                    //  Element 1 is reserved for the target of the repeat (and needs setting here)
-                    onRepeat.params[1] = tweenData.target;
-    
-                    onRepeat.func.apply(onRepeat.scope, onRepeat.params);
-                }
+                this.dispatchTweenDataEvent(Events.TWEEN_REPEAT, tween.callbacks.onRepeat, tweenData);
 
                 return TWEEN_CONST.PLAYING_FORWARD;
             }
@@ -1369,7 +1391,7 @@ var Tween = new Class({
                 tweenData.target.toggleFlipY();
             }
 
-            tweenData.end = tweenData.getEndValue(tweenData.target, tweenData.key, tweenData.start);
+            tweenData.end = tweenData.getEndValue(tweenData.target, tweenData.key, tweenData.start, tweenData.index, tween.totalTargets, tween);
 
             //  Delay?
             if (tweenData.repeatDelay > 0)
@@ -1384,17 +1406,7 @@ var Tween = new Class({
             }
             else
             {
-                this.emit(Events.TWEEN_REPEAT, this, tweenData.key, tweenData.target);
-
-                var onRepeat = tween.callbacks.onRepeat;
-    
-                if (onRepeat)
-                {
-                    //  Element 1 is reserved for the target of the repeat (and needs setting here)
-                    onRepeat.params[1] = tweenData.target;
-    
-                    onRepeat.func.apply(onRepeat.scope, onRepeat.params);
-                }
+                this.dispatchTweenDataEvent(Events.TWEEN_REPEAT, tween.callbacks.onRepeat, tweenData);
 
                 return TWEEN_CONST.PLAYING_FORWARD;
             }
@@ -1413,18 +1425,20 @@ var Tween = new Class({
      *
      * @param {Phaser.Tweens.Tween} tween - The Tween to update.
      * @param {Phaser.Types.Tweens.TweenDataConfig} tweenData - The TweenData property to update.
-     * @param {number} delta - Either a value in ms, or 1 if Tween.useFrames is true
+     * @param {number} delta - Either a value in ms, or 1 if Tween.useFrames is true.
      *
      * @return {boolean} True if the tween is not complete (e.g., playing), or false if the tween is complete.
      */
     updateTweenData: function (tween, tweenData, delta)
     {
+        var target = tweenData.target;
+
         switch (tweenData.state)
         {
             case TWEEN_CONST.PLAYING_FORWARD:
             case TWEEN_CONST.PLAYING_BACKWARD:
 
-                if (!tweenData.target)
+                if (!target)
                 {
                     tweenData.state = TWEEN_CONST.COMPLETE;
                     break;
@@ -1445,39 +1459,16 @@ var Tween = new Class({
                 var forward = (tweenData.state === TWEEN_CONST.PLAYING_FORWARD);
                 var progress = elapsed / duration;
 
-                var v;
-
-                if (forward)
-                {
-                    v = tweenData.ease(progress);
-                }
-                else
-                {
-                    v = tweenData.ease(1 - progress);
-                }
-
-                tweenData.current = tweenData.start + ((tweenData.end - tweenData.start) * v);
-
-                tweenData.target[tweenData.key] = tweenData.current;
-
                 tweenData.elapsed = elapsed;
                 tweenData.progress = progress;
-
-                this.emit(Events.TWEEN_UPDATE, this, tweenData.key, tweenData.target);
-
-                var onUpdate = tween.callbacks.onUpdate;
-
-                if (onUpdate)
-                {
-                    onUpdate.params[1] = tweenData.target;
-
-                    onUpdate.func.apply(onUpdate.scope, onUpdate.params);
-                }
 
                 if (progress === 1)
                 {
                     if (forward)
                     {
+                        tweenData.current = tweenData.end;
+                        target[tweenData.key] = tweenData.end;
+
                         if (tweenData.hold > 0)
                         {
                             tweenData.elapsed = tweenData.hold - diff;
@@ -1491,9 +1482,22 @@ var Tween = new Class({
                     }
                     else
                     {
+                        tweenData.current = tweenData.start;
+                        target[tweenData.key] = tweenData.start;
+
                         tweenData.state = this.setStateFromStart(tween, tweenData, diff);
                     }
                 }
+                else
+                {
+                    var v = (forward) ? tweenData.ease(progress) : tweenData.ease(1 - progress);
+
+                    tweenData.current = tweenData.start + ((tweenData.end - tweenData.start) * v);
+
+                    target[tweenData.key] = tweenData.current;
+                }
+
+                this.dispatchTweenDataEvent(Events.TWEEN_UPDATE, tween.callbacks.onUpdate, tweenData);
 
                 break;
 
@@ -1516,21 +1520,11 @@ var Tween = new Class({
 
                 if (tweenData.elapsed <= 0)
                 {
-                    this.emit(Events.TWEEN_REPEAT, this, tweenData.key, tweenData.target);
-
-                    var onRepeat = tween.callbacks.onRepeat;
-        
-                    if (onRepeat)
-                    {
-                        //  Element 1 is reserved for the target of the repeat (and needs setting here)
-                        onRepeat.params[1] = tweenData.target;
-        
-                        onRepeat.func.apply(onRepeat.scope, onRepeat.params);
-                    }
-
                     tweenData.elapsed = Math.abs(tweenData.elapsed);
 
                     tweenData.state = TWEEN_CONST.PLAYING_FORWARD;
+
+                    this.dispatchTweenDataEvent(Events.TWEEN_REPEAT, tween.callbacks.onRepeat, tweenData);
                 }
 
                 break;
@@ -1548,15 +1542,15 @@ var Tween = new Class({
 
             case TWEEN_CONST.PENDING_RENDER:
 
-                if (tweenData.target)
+                if (target)
                 {
-                    tweenData.start = tweenData.getStartValue(tweenData.target, tweenData.key, tweenData.target[tweenData.key]);
+                    tweenData.start = tweenData.getStartValue(target, tweenData.key, target[tweenData.key], tweenData.index, tween.totalTargets, tween);
 
-                    tweenData.end = tweenData.getEndValue(tweenData.target, tweenData.key, tweenData.start);
+                    tweenData.end = tweenData.getEndValue(target, tweenData.key, tweenData.start, tweenData.index, tween.totalTargets, tween);
 
                     tweenData.current = tweenData.start;
 
-                    tweenData.target[tweenData.key] = tweenData.start;
+                    target[tweenData.key] = tweenData.start;
 
                     tweenData.state = TWEEN_CONST.PLAYING_FORWARD;
                 }